+++ tests/test_home.py
@@ -14,7 +14,10 @@ from homematicip.device import Device
 
 import json
 from datetime import datetime, timedelta, timezone
-from homematicip_demo.helper import fake_home_download_configuration, no_ssl_verification
+from homematicip_demo.helper import (
+    fake_home_download_configuration,
+    no_ssl_verification,
+)
 from conftest import utc_offset